# db/model/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.22.0

package db

import (
	"database/sql"
)

type Task struct {
	ID          int64          `json:"id"`
	Username    string         `json:"username"`
	Title       sql.NullString `json:"title"`
	Description sql.NullString `json:"description"`
	DueDate     sql.NullTime   `json:"due_date"`
	Priority    sql.NullInt32  `json:"priority"`
	CreatedAt   sql.NullTime   `json:"created_at"`
}

type User struct {
	Username          string       `json:"username"`
	FullName          string       `json:"full_name"`
	HashedPassword    string       `json:"hashed_password"`
	Email             string       `json:"email"`
	PasswordChangedAt sql.NullTime `json:"password_changed_at"`
	CreatedAt         sql.NullTime `json:"created_at"`
}